[create]:create user MVC
